Refactored and fixed simple components
