+++ src/components/ReceiveScreen/ReceiveScreen.js
@@ -6,23 +6,23 @@ import {compose} from 'redux'
 import {View} from 'react-native'
 
 import Screen from '../../components/Screen'
-import OwnAddressesList from './OwnAddressesList'
+import ReceiveAddressesList from './ReceiveAddressesList'
 
 import styles from './ReceiveScreen.style'
 
 type Props = {
-  ownAddresses: Array<string>,
+  receiveAddresses: Array<string>,
 };
 
 
 class ReceiveScreen extends Component<Props> {
   render() {
-    const {ownAddresses} = this.props
+    const {receiveAddresses} = this.props
 
     return (
       <View style={styles.root}>
         <Screen scroll>
-          <OwnAddressesList ownAddresses={ownAddresses} />
+          <ReceiveAddressesList receiveAddresses={receiveAddresses} />
         </Screen>
       </View>
     )
@@ -31,6 +31,6 @@ class ReceiveScreen extends Component<Props> {
 
 export default compose(
   connect((state) => ({
-    ownAddresses: state.ownAddresses,
+    receiveAddresses: state.receiveAddresses,
   })),
 )(ReceiveScreen)
