+++ src/components/ReceiveScreen/ReceiveScreenNavigator.js
@@ -3,14 +3,14 @@
 import {createStackNavigator} from 'react-navigation'
 import ReceiveScreen from './ReceiveScreen'
 
-const SEND_ROUTES = {
+const RECEIVE_ROUTES = {
   MAIN: 'main',
 }
 
 const ReceiveScreenNavigator = createStackNavigator({
-  [SEND_ROUTES.MAIN]: ReceiveScreen,
+  [RECEIVE_ROUTES.MAIN]: ReceiveScreen,
 }, {
-  initialRouteName: SEND_ROUTES.MAIN,
+  initialRouteName: RECEIVE_ROUTES.MAIN,
   navigationOptions: ({navigation}) => ({
     title: 'Receive ADA',
   }),
